PMM-10503 change view based on form and call API
