+++ public/app/percona/backup/components/ScheduledBackups/ScheduledBackups.service.ts
@@ -2,8 +2,6 @@ import { CancelToken } from 'axios';
 
 import { api } from 'app/percona/shared/helpers/api';
 
-import { BackupMode, DataModel } from '../../Backup.types';
-
 import { ScheduledBackup, ScheduledBackupResponse } from './ScheduledBackups.types';
 
 const BASE_URL = '/v1/management/backup/Backups';
@@ -57,58 +55,10 @@ export const ScheduledBackupsService = {
       })
     );
   },
-  async schedule(
-    serviceId: string,
-    locationId: string,
-    cronExpression: string,
-    name: string,
-    description: string,
-    retryInterval: string,
-    retryTimes: number,
-    retention: number,
-    enabled: boolean,
-    mode: BackupMode,
-    dataModel: DataModel
-  ) {
-    return api.post(`${BASE_URL}/Schedule`, {
-      service_id: serviceId,
-      location_id: locationId,
-      cron_expression: cronExpression,
-      name,
-      description,
-      retry_interval: retryInterval,
-      retries: retryTimes,
-      enabled: !!enabled,
-      retention,
-      mode,
-      data_model: dataModel,
-    });
-  },
   async toggle(id: string, enabled: boolean) {
     return api.post(`${BASE_URL}/ChangeScheduled`, { scheduled_backup_id: id, enabled });
   },
   async delete(id: string) {
     return api.post(`${BASE_URL}/RemoveScheduled`, { scheduled_backup_id: id });
-  },
-  async change(
-    id: string,
-    enabled: boolean,
-    cronExpression: string,
-    name: string,
-    description: string,
-    retryInterval: string,
-    retryTimes: number,
-    retention: number
-  ) {
-    return api.post(`${BASE_URL}/ChangeScheduled`, {
-      scheduled_backup_id: id,
-      enabled,
-      cron_expression: cronExpression,
-      name,
-      description,
-      retry_interval: retryInterval,
-      retries: retryTimes,
-      retention,
-    });
   },
 };
