+++ public/app/percona/backup/components/ScheduledBackups/ScheduledBackups.tsx
@@ -18,10 +18,9 @@ import { usePerconaNavModel } from 'app/percona/shared/components/hooks/perconaN
 import { DATABASE_LABELS } from 'app/percona/shared/core';
 import { getPerconaSettingFlag } from 'app/percona/shared/core/selectors';
 import { isApiCancelError } from 'app/percona/shared/helpers/api';
-import { getCronStringFromValues } from 'app/percona/shared/helpers/cron/cron';
 
 import { Messages } from '../../Backup.messages';
-import { RetryMode } from '../../Backup.types';
+import { BackupService } from '../../Backup.service';
 import { formatBackupMode } from '../../Backup.utils';
 import { AddBackupModal } from '../AddBackupModal';
 import { AddBackupFormProps } from '../AddBackupModal/AddBackupModal.types';
@@ -85,12 +84,11 @@ export const ScheduledBackups: FC = () => {
         retryInterval,
         retryTimes,
         mode,
-        dataModel,
       } = backup;
       const newName = `${Messages.scheduledBackups.copyOf} ${name}`;
       setActionPending(true);
       try {
-        await ScheduledBackupsService.schedule(
+        await BackupService.scheduleBackup(
           serviceId,
           locationId,
           cronExpression,
@@ -101,7 +99,6 @@ export const ScheduledBackups: FC = () => {
           retention,
           false,
           mode,
-          dataModel
         );
         getData();
       } catch (e) {
@@ -200,67 +197,13 @@ export const ScheduledBackups: FC = () => {
     setBackupModalVisible(false);
   };
 
-  const handleBackup = async (backup: AddBackupFormProps) => {
-    const {
-      id,
-      service,
-      location,
-      period,
-      month,
-      day,
-      weekDay,
-      startHour,
-      startMinute,
-      backupName,
-      description,
-      retryMode,
-      retryInterval,
-      retryTimes,
-      active,
-      retention,
-      mode,
-      dataModel,
-    } = backup;
+  const handleBackup = async (values: AddBackupFormProps) => {
+    const { id, backupName } = values;
     try {
-      const cronExpression = getCronStringFromValues(
-        period!.value!,
-        month!.map((m) => m.value!),
-        day!.map((m) => m.value!),
-        weekDay!.map((m) => m.value!),
-        startHour!.map((m) => m.value!),
-        startMinute!.map((m) => m.value!)
-      );
-      const strRetryInterval = `${retryInterval}s`;
-      let resultRetryTimes = retryMode === RetryMode.MANUAL ? 0 : retryTimes;
-
-      if (id) {
-        await ScheduledBackupsService.change(
-          id,
-          active!,
-          cronExpression,
-          backupName,
-          description,
-          strRetryInterval,
-          resultRetryTimes!,
-          retention!
-        );
-        appEvents.emit(AppEvents.alertSuccess, [Messages.scheduledBackups.getEditSuccess(backupName)]);
-      } else {
-        await ScheduledBackupsService.schedule(
-          service!.value?.id!,
-          location!.value!,
-          cronExpression,
-          backupName,
-          description,
-          strRetryInterval,
-          resultRetryTimes!,
-          retention!,
-          active!,
-          mode,
-          dataModel
-        );
-        appEvents.emit(AppEvents.alertSuccess, [Messages.scheduledBackups.addSuccess]);
-      }
+      await BackupService.backup(values);
+      appEvents.emit(AppEvents.alertSuccess, [
+        id ? Messages.scheduledBackups.getEditSuccess(backupName) : Messages.scheduledBackups.addSuccess,
+      ]);
       setBackupModalVisible(false);
       setSelectedBackup(null);
       getData();
